created class MovieSearch
